VariableHood, SetHoodState command, distance, smartdashboard
SetHoodState sets the hood to one of 4 (up to 8 if we want to do that later) positions

distance logic is super scuffed right now (I just used if elses, we might want to change later, I just thought we should try to get it ready for data point testing today)

smartdashboard as a called method in Robot.java is done. I will make it as an object within a command in the next commit and we can comment out the cringier one.