made the graph another layer deep
The graph is backed by an array of adjacency matrices. The matrix is now setup to be n x n x n where n is the number of items that have been registered. The point of this is that we can now get a more accurate estimate of the field with an approach that isn't necessarily from the x axis. This helps with our AngleFromTo method in our node class. The space complexity of the graph is now O(n^3) however we also have a few array lists and maps in there so it is very large